add detail/iterators/iteration_proxy.hpp
